Rejiggle negrcpdjacs at the soln points.
For the viscous terms we will need J.I.T which, as the S-matrices
are defined as S = |J|J.I requires us to have 1/|J| (rcpdjacs)
available since J.I = S/|J|.  However, prior to this commit we
have -1/|J| which is not suitable.  This commit makes 1/|J|
available (as opposed to -1/|J|).