Update docs and conventionalise heading level characters.
The characters that define the level of headings in PyFR's documentation
have been updated to follow the Sphinx convention for Python code:
* (under and overline) for chapters
= (underline) for sections
- (underline) for subsections
^ (underline) for subsubsections
" (underline) for paragraphs

Note:
The "Developer's Guide" is expected to be made into a seperate document,
and has heading levels defined as such. It is currently featured as a
chapter though, which leads to a temporary exception to the above rule.