Add a cut-down DOPRI5 integrator.
This is a fifth order scheme with six stages.  Although currently
unimplemented the scheme supports both step size control and dense
output.  Seven registers of storage are required (compared to five
for RK4) although this will increase to eight when step size
control is implemented.  (And it is possible to cut the scheme
down to six if such control is not required.)  Stability should
be excellent.