Reduce the memory requirements of RK4 from 5 to 3 registers.
Although a full-fat 4th order method with reasonable coefficients
by employing a few tricks it is possible to greatly reduce the
memory requirements compared with a naïve appproach.  This is
close to the two registers required by a low-storage method.

The performance impact from this (due to the extra BLAS calls) has
been measured at < 1%.  The memory saving is easily over 10%.