Further restructure the backends.
Pointwise kernels are now specified in terms of a minimal,
templated, domain specific language.  Each backend has a
pointwise generator module which translates the generic
description of a pointwise kernel into the underlying language
of the backend.  This allows a common set of pointwise
kernels to be used for all pointwise operations.

All kernels except for the HLL Riemann solver have been ported.
However, a large amount of cleanup, both of the generators
themselves, and of the surrounding legacy code are still required.