Implemented the parallel vtu file writer.
This commit implements the parallel vtu file writer for visualization
with paraview. The writer is invoked whtn the output file to be written
has a .pvtu extension.

The commit also implements the feature where the size of an array can
be ascertained without actually loading the array, when the underlying
file format supports it, as in the case of the new file format based on
HDF5.