Implementation of space and time dependent bc.
This implements the framework for a spatially and temporally
varying boundary condition. As an extension to the implementation
the sub-in-frv, sup-in-fa are now capable of taking in equations
in terms of x, y, z and t. Also, for the ease of usage, the
constants defined in the constants section are also available
in the equations.