Rework artificial viscosity.
Support is now present at the BaseAdvectionDiffusion level as
opposed to just Navier-Stokes.  Further, sensing in the case of
Navier-Stokes is now based on the density field as opposed to the
entropy.  Additionally, artificial viscosity is treated
separately from the physical viscosity (if any).  For Navier-Stokes
this means that the density field now has viscosity applied to it.