Where possible explicitly finalise MPI.
This should cut down on the number of (harmless) MPI related
segfaults which occur on exit.

Note that at the point MPI_Finalize is invoked there are indeed
still some MPI requests extant in the solver.  However, when
these are garbage collected mpi4py first checks the status of MPI
before releasing them.