Add support for post-actions in plugins.
A post-action is a command (either blocking or non-blocking) to
be invoked on the root rank after a plugin has completed.
Support for post-actions currently exists in the writer plugin
with the syntax being:

[soln-plugin-writer]
...
post-action = myprog "{mesh}" "{soln}" {t}
post-action-mode = non-blocking

which will pass the mesh file name, the just-written solution
file name and the current simulation time to myprog.  The
simulation will then continue execution while the plugin
executes.  However, myprog will not be allowed to execute again
until the current invocation has finished (and hence is myprog
is extremely long-lived then there is the possibility of it
blocking the the solver).