Cleanup the rank allocation code.
We now save the rank allocation inside of the stats file as a
comma separated list of partition numbers.  Hence, the first
item is the partition assigned to the first MPI rank.

A random rank allocation strategy has also been added.  This is
useful for assessing the impact of the cluster topoplogy on a
simulation.