Plugin: add stop conditions during runtime
This plugin adds certain mechanisms to stop simulations
writing all solution data before exiting

`max-phystime`:  if the physical time of the simulations
is higher than this parameter, all solution data will be
written to disk and the simulation will end.

`max-walltime`: if the wall clock time measured from the
start of the simulation is bigger than this parameter,
all solution data will be written to disk and the simulation
will end.

`stop-file`: if a file with the name associated to this parameter
is found, all solution data will be written to disk and the
simulation will end.