Writer: added high-order vtu output
This commit adds a new writer which outputs true
high-order VTK cells using the vtkLagrange* cell types.

The main benefits of this new writer are:

1) Appropriate visualization of non-linear cells within
Paraview.

2) The user may also modify the parameter
`Nonlinear subdivision Level` to increase or reduce the
interpolation and shape basis order used by VTK
or Paraview to represent the solution and cells' shape.

3) There is no need to output gradient data from
PyFR as VTK will use the same interpolation basis
than the one in PyFR, provided that the number of
subdivisions is consistent with the interpolation
order of the simulations.

4) File sizes are a smaller (40% smaller in some cases)
as less cells are output and due to the fact that
the connectivity arrays have less entries.

The main downsides of this implementation is that
VTK is not well-equiped to deal with high-order
cells and that Paraview becomes slower.
One can alleviate a bit the latter issue by
cleaning the grid.

To use this writer one needs to add the argument
`-t vtkho` to the export command

WARNING: `pyr` type elements are not compatible as
there exists no `vtkLagrangePyramid` element in VTK.
It would be possible to use `vtkQuadraturePyramid`
for a quadratic `pyr` representation in VTK.
However, for the moment this writer will
subdivide `pyr` elements if found in the mesh.

INFO: the `nodemaps.py` file includes the mapping
of nodes from VTK8-9 to PyFR and viceversa.
A small note has been added to explain the
importance of using the VTK8 node connectivity
instead of that of VTK9