integrate: choose quadrature degree.
Allow to choose the quadrature degree used to evaluate the
expressions in the plugin integrate. To do so, the token
quad-deg needs to be added to the input file in the integrate
plugin section. The default quadrature type is taken from that used
to set the solution points. If the user adds
quad-pts-{etype} = quad_name to the integrate plugin
section, then the the quad_name quadrature will be used.

Example:

[soln-plugin-integrate]
nsteps = 10
quad-deg = 10 ; quadrature degree
quad-pts-tet = witherden-vincent ; quadrature type
file = integrate
int-rho = rho